refactoring checkpoint 2
