+++ src/communication_generic.h
@@ -1,0 +1,31 @@
+#ifndef COMMUNICATION_PRECISION_H
+#define COMMUNICATION_PRECISION_H
+#ifndef IMPORT_FROM_EXTERN_C
+#include <mpi.h>
+#endif
+#include "complex_types_PRECISION.h"
+
+typedef struct
+{
+    int length[8], *boundary_table[8], max_length[4],
+        comm_start[8], in_use[8], offset, comm,
+        num_even_boundary_sites[8], num_odd_boundary_sites[8],
+        num_boundary_sites[8];
+#ifdef CUDA_OPT
+    int *boundary_table_gpu[8];
+    cuda_vector_PRECISION buffer_gpu[8];
+#endif
+    vector_PRECISION buffer[8];
+    MPI_Request sreqs[8], rreqs[8];
+} comm_PRECISION_struct;
+
+typedef struct
+{
+    int ilde, dist_local_lattice[4], dist_inner_lattice_sites,
+        *permutation, *gather_list, gather_list_length;
+    vector_PRECISION buffer, transfer_buffer;
+    MPI_Request *reqs;
+    MPI_Group level_comm_group;
+    MPI_Comm level_comm;
+} gathering_PRECISION_struct;
+#endif // COMMUNICATION_PRECISION_H
